Refactoring and cleanup in preparation for a feature release of this gem.
This will include things like:

* Easy to understand code cleanup.
* Default test suites available within the gem, executed when no
* parameter is given.
* Consistent namespacing.
* Cleanup the runner class, plus the code base in there.
* The setup script is now dinamically getting all relevent data and
  running the test in there. (rvm required)

And many more.
