+++ .gitignore
@@ -1,2 +1,4 @@
+*.lock
 *.gem
 config.txt
+workspace/
