--- bin/lsit-suite
@@ -1,24 +1,0 @@
-#!/usr/bin/env ruby
-
-# encoding: utf-8
-
-$LOAD_PATH << "."
-
-require 'lsit'
-
-@debug = !!ENV['DEBUG']
-
-if ARGV.size < 1 or ARGV.size > 2
-  $stderr.puts("usage: ruby suite.rb [suite file] [logstash path]")
-  exit(1)
-end
-
-debug  = !!ENV['DEBUG']
-
-logstash_home = ENV['LOGSTASH_HOME'] || 'logstash'
-install_path  = ARGV.size > 1 ? ARGV[1] : logstash_home
-
-runner = LSit::Executor::Suite.new(ARGV[0], install_path)
-runner.config = '.lsit' if File.exist?('.lsit.yml')
-lines  = runner.execute(debug)
-puts lines.join("\n")
