+++ bin/lsperfm
@@ -1,0 +1,9 @@
+#!/usr/bin/env ruby
+
+# encoding: utf-8
+
+$LOAD_PATH << "."
+
+require 'lsperfm'
+
+LogStash::PerfM.invoke
