--- lib/lsit.rb
@@ -1,9 +1,0 @@
-require 'yaml'
-require 'lsit/reporter'
-require 'lsit/stats'
-require 'lsit/run'
-require 'lsit/suite'
-
-module LSit
-
-end
