--- lib/lsit/version.rb
@@ -1,3 +1,0 @@
-module LSit
-  VERSION='0.1.0'
-end
