+++ lib/lsperfm/defaults/config/json_inout_codec.conf
@@ -1,0 +1,11 @@
+input {
+  stdin { codec => "json_lines" }
+}
+
+filter {
+  clone {}
+}
+
+output {
+  stdout { codec => json_lines }
+}
