+++ lib/lsperfm/defaults/config/json_inout_filter.conf
@@ -1,0 +1,11 @@
+input {
+  stdin {}
+}
+
+filter {
+  json { source => "message" }
+}
+
+output {
+  stdout { codec => json_lines }
+}
