+++ lib/lsperfm/defaults/config/simple.conf
@@ -1,0 +1,11 @@
+input {
+  stdin {}
+}
+
+filter {
+  clone {}
+}
+
+output {
+  stdout { codec => line }
+}
