+++ lib/lsperfm/defaults/config/simple_grok.conf
@@ -1,0 +1,13 @@
+input {
+	stdin { type => "apache" }
+}
+
+filter {
+	grok {
+		match => {"message" => "%{COMBINEDAPACHELOG}"}
+	}
+}
+
+output {
+	stdout { codec => line }
+}
