+++ lib/lsperfm/defaults/config/simple_json_out.conf
@@ -1,0 +1,11 @@
+input {
+  stdin {}
+}
+
+filter {
+  clone {}
+}
+
+output {
+  stdout { codec => json_lines }
+}
