+++ lib/lsperfm/defaults/input/simple_10.txt
@@ -1,0 +1,10 @@
+test 01
+test 02
+test 03
+test 04
+test 05
+test 06
+test 07
+test 08
+test 09
+test 10
