+++ lib/lsperfm/version.rb
@@ -1,0 +1,5 @@
+module LogStash
+  module PerfM
+    VERSION='0.1.0'
+  end
+end
