--- monit.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-parallel --no-notice --col-sep ' ' --wd . --progress -a config.txt ./scripts/setup.sh
