+++ scripts/setup.sh
@@ -2,29 +2,65 @@
 
 [[ -s "$HOME/.rvm/scripts/rvm" ]] && source "$HOME/.rvm/scripts/rvm"
 
-function run() {
-  ORIGIN=`pwd`
-  echo "Updatting the repository $1"
-  cd  $1 #jumping into the repository
-  git checkout -q $2 #checkout into the branch
-  git pull # update the repo
-
-  echo "Dependency installation in $1 $ORIGIN"
-  cd $ORIGIN
-  ruby install_deps.rb $1
-
-  echo "Basic performance report"
-  ruby suite.rb suite/basic_performance_quick.rb $1 > "$3"
-  echo "done"
-}
-
 TODAY=`date +%s`
 
-DIR=$1
-BRANCH=$2
-REPORT="../basic_performance_report-$2_$TODAY.csv"
+VERSIONS=( 1.4.2 1.5.0 1.5.1 1.5.2 )
+DEST_DIR="workspace"
 
-echo "run performance setup"
-run $DIR $BRANCH $REPORT
-echo "done"
+rm -rf $DEST_DIR
+mkdir -p $DEST_DIR
 
+## Download released versions.
+
+for VERSION in "${VERSIONS[@]}"
+do
+
+  FILENAME="logstash-$VERSION.tar.gz"
+  SOURCE_FILE="$DEST_DIR/$FILENAME"
+  DOWNLOAD_URL="https://download.elasticsearch.org/logstash/logstash/$FILENAME"
+
+  if [ ! -f $SOURCE_FILE ]; then
+	  echo "Downloading $DOWNLOAD_URL"
+    wget $DOWNLOAD_URL -O $SOURCE_FILE
+    tar -xvzf $SOURCE_FILE -C $DEST_DIR
+  fi
+done
+
+
+## Bootstrapping the environment
+rvm use --create jruby-1.7.20@logstash-perfm
+gem install logstash-perftool
+
+cd $DEST_DIR
+## Run the report for each download package
+for VERSION in "${VERSIONS[@]}"
+do
+ cd "logstash-$VERSION"
+ rvm use jruby-1.7.20@logstash-perfm
+ lsperfm > "../logstash-$VERSION-$TODAY.log"
+ cd ..
+done
+
+exit 0
+
+## Setup the current codebase
+rm -rf "logstash"
+git clone git@github.com:elastic/logstash.git
+echo "jruby-1.7.20"   > "logstash/.ruby-version"
+echo "logstash-perfm" > "logstash/.ruby-gemset"
+
+cd logstash
+
+## Running the benchmarks for each logstash repository branch
+## of interest.
+BRANCHES=( master 1.5 )
+for BRANCH in "${BRANCHES[@]}"
+do
+  git checkout Gemfile
+  git checkout Gemfile.jruby-1.9.lock
+  git checkout $BRANCH
+  rm -rf vendor
+  rake bootstrap
+  lsperfm-deps
+  lsperfm > "../logstash-branch-$BRANCH-$TODAY.log"
+done
