+++ spec/lib/suite_spec.rb
@@ -1,15 +1,15 @@
 require 'spec_helper'
 
-describe LSit::Executor::Suite do
+describe LogStash::PerfM::Core do
 
   let(:config)        { 'spec/fixtures/config.yml' }
   let(:logstash_home) { '.' }
   let(:suite_def)     { 'spec/fixtures/basic_suite.rb' }
   let(:serial_runner) { double('DummySerialRunner') }
-  let(:runner)        { Runner }
+  let(:runner)        { LogStash::PerfM::Runner }
 
   let(:run_outcome)   { { :percentile => [2000] , :elapsed => 100, :events_count => 3000, :start_time => 12 } }
-  subject(:manager) { LSit::Executor::Suite.new(suite_def, logstash_home, config, runner) }
+  subject(:manager) { LogStash::PerfM::Core.new(suite_def, logstash_home, config, runner) }
 
   context "with a valid configuration" do
     before(:each) do
@@ -20,7 +20,7 @@ describe LSit::Executor::Suite do
 
       it "run each test case in a serial maner" do
         expect(runner).to receive(:new).with("spec/fixtures/simple.conf", false, logstash_home).twice { serial_runner }
-        manager.execute
+        manager.run
       end
 
     end
@@ -32,7 +32,7 @@ describe LSit::Executor::Suite do
       it "run each test case as expected" do
         expect(runner).to receive(:read_input_file).with('./simple_10.txt').twice { [] }
         expect(runner).to receive(:new).with("./simple.conf", false, logstash_home).twice { serial_runner }
-        manager.execute
+        manager.run
       end
 
     end
@@ -44,7 +44,7 @@ describe LSit::Executor::Suite do
 
     it "run each test case as expected" do
       expect(runner).to receive(:new).with("spec/wrong_fixture/simple.conf", false, logstash_home).once { serial_runner }
-      expect { manager.execute }.to raise_error(LSit::ConfigException)
+      expect { manager.run }.to raise_error(LogStash::PerfM::ConfigException)
     end
   end
