+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@ ROOT = File.expand_path(File.join(File.dirname(__FILE__), '..'))
 $LOAD_PATH.unshift File.join(ROOT, 'lib')
 $LOAD_PATH.unshift File.join(ROOT, 'spec')
 
-require 'lsit'
+require 'lsperfm'
 
 def load_fixture(name)
   IO.read("spec/fixtures/#{name}")
